开关充电。TODO：查询时更新实时电量

+++ src/main/java/com/bupt/charger/response/CarStatusResponse.java
@@ -18,5 +18,5 @@ public class CarStatusResponse extends Resp {
     private String queueNum;
 
     @JsonProperty("request_time")
-    private int requestTime = -1;
+    private long requestTime = -1;
 }
